refactor(repo): skill-repository → assignment + row-mapper 분리
PR #44/#45 의 패턴 follow-up. skill-repository.ts (528 lines) 의 assignment
영역을 별도 module 로 추출.

# 동기
- max-lines warning 해소 (528 → 499 lines, eslint warn 사라짐)
- 책임 분리 명확화:
  - SkillRepository (499): skill CRUD + 검색 + draft + 시스템 upsert + status
  - SkillAssignmentRepository (138): agent_skill_assignments 테이블 전용
    (assign/remove/getSkillsForAgent/getSkillIdsForAgent + user 변형 4개)
  - skill-row-mapper.ts (49): rowToSkill + toStringValue/toBooleanValue/
    toDateValue (양쪽 repo 공용 helper)

# Breaking change 0 — facade 패턴
- SkillRepository 의 8 assignment 메서드는 thin facade 로 유지:
  `private get assignments() { return new SkillAssignmentRepository(this.pool); }`
- 기존 사용자 (skill-seeder, system-prompt, skill-manager, agents.routes,
  skills.routes) 의 호출 코드 영향 0
- 직접 SkillAssignmentRepository import 도 가능 — 신규 코드는 책임 분리 권장

# 검증
- tsc 0 / lint 0 errors / 0 warnings (모든 repositories file)
- 전체 회귀: 97 suites / 1604 tests / 0 failed (동일)
- repository 디렉토리 max-lines warning 완전 해소
